Add mpris to web
