+++ disconnect-module/src/main.rs
@@ -1,15 +1,13 @@
 use cli_module::{
     DelayArgs, SharedArgs, SharedCommands, create_player, default_audio_cache,
-    default_audio_quality, get_client, handle_shared_commands, spawn_clean_up,
+    default_audio_quality, error_exit, get_client, handle_shared_commands, spawn_clean_up,
 };
 use disconnect_module::DisconnectClientConfig;
 use std::sync::Arc;
 use tokio::sync::{broadcast, mpsc, watch};
 
 use clap::Parser;
-use player_module::{
-    AppResult, database::Database, error::Error, notification::NotificationBroadcast,
-};
+use player_module::{AppResult, database::Database, notification::NotificationBroadcast};
 
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
@@ -137,9 +135,4 @@ pub async fn run() -> AppResult<()> {
     player.player_loop(exit_receiver).await?;
 
     Ok(())
-}
-
-fn error_exit(error: Error) {
-    eprintln!("{error}");
-    std::process::exit(1);
 }
